Remove namespace of LinksPublisher and PublishTags
Both these service classes were nested under `TagImporter`. Although that
particular feature uses those classes, there are other features that use them
too (e.g. `TagMigration`).

This commit deletes the namespaces to avoid confusion.
