+++ app/controllers/tag_migrations_controller.rb
@@ -50,7 +50,7 @@ class TagMigrationsController < ApplicationController
   end
 
   def publish_tags
-    TagImporter::PublishTags.new(tag_migration, user: current_user).run
+    PublishTags.new(tag_migration, user: current_user).run
 
     redirect_to(
       tag_migration,
